+++ CHANGELOG.md
@@ -13,6 +13,10 @@ v0.4.0 (in development)
   the directory in which daemail was started, not the directory specified with
   `--chdir`
 
+- No more than one line ending sequence will be removed from
+  `--smtp-password-file` (for the unlikely case in which the password ends with
+  a newline)
+
 
 v0.3.0 (2016-08-28)
 -------------------